+++ src/main/kotlin/com/loadingbyte/cinecred/project/Project.kt
@@ -255,6 +255,7 @@ class Block(
 typealias StyledString = List<Pair<String, LetterStyle>>
 
 sealed class BodyElement {
+    class Nil(val sty: LetterStyle) : BodyElement()
     class Str(val str: StyledString) : BodyElement()
     class Pic(val pic: Picture) : BodyElement()
 }